backend/requirements.txt
absl-py==1.4.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==23.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.8.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cleo==0.8.1
clerk==0.1.0
click==8.1.7
clikit==0.6.2
colorama==0.4.6
contourpy==1.2.1
crashtest==0.3.1
cryptography==36.0.2
cycler==0.12.1
databases==0.9.0
DateTime==5.5
Deprecated==1.2.14
dnspython==2.7.0
dotty-dict==1.3.1
email_validator==2.2.0
exceptionite==2.2.5
Faker==13.16.0
fastapi==0.115.14
filelock==3.15.4
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.53.0
fsspec==2024.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.58.0
h11==0.16.0
h5py==3.9.0
hashids==1.3.1
hfilesize==0.1.0
hupper==1.10.3
idna==3.4
inflection==0.3.1
iniconfig==2.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lxml==5.2.2
Markdown==3.4.4
MarkupSafe==2.1.3
masonite==4.20.2
masonite-orm==2.24.0
matplotlib==3.9.0
mkl==2021.4.0
mock==5.2.0
mpmath==1.3.0
neat-python==0.92
networkx==3.3
nnfs==0.5.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.2.2
pastel==0.2.1
pendulum==2.1.2
phonenumbers==8.13.55
Pillow==10.0.0
pluggy==1.6.0
protobuf==4.24.3
psycopg2-binary==2.9.10
psycopg2-binary==2.9.10
pyarrow==16.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybaseball==2.2.7
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.0
PyGithub==2.3.0
PyJWT==2.4.0
pylev==1.4.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
pytz==2024.1
pytzdata==2020.1
redis==5.0.4
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.11.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spotipy==2.23.0
SQLAlchemy==2.0.41
starlette==0.46.2
stripe==12.2.0
sympy==1.13.1
tbb==2021.13.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.5.0
tldextract==2.2.3
tqdm==4.66.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.16
uvicorn==0.35.0
watchdog==4.0.0
Werkzeug==3.0.2
whitenoise==5.2.0
wrapt==1.15.0
zope.interface==7.2